update package after subproject rename
