+++ modules/mockk-core/src/commonMain/kotlin/io/mockk/core/ValueClassSupport.kt
@@ -1,4 +1,4 @@
-package io.mockk.platform
+package io.mockk.core
 
 import kotlin.reflect.KClass
